career_chatbot/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
spacy==3.5.0
joblib==1.3.1
python-dotenv==1.0.0
werkzeug==2.3.0
PyMySQL==1.1.0
mysql-connector-python==8.0.33
groq==0.4.2
